Refactor client, server, and shared code
