+++ packages/usellm/src/shared/utils.ts
@@ -1,24 +1,8 @@
-export interface OpenAIMessage {
-  content: string;
-  role: string;
-  user?: string;
-}
-
-export type ChatStreamCallback = (result: {
-  message: OpenAIMessage;
-  isFirst: boolean;
-  isLast: boolean;
-}) => void;
-
-export interface LLMChatResult {
-  message: OpenAIMessage;
-}
-
-export interface ScoreEmbeddingsOptions {
-  embeddings: Array<Array<number>>;
-  query: number[];
-  top?: number;
-}
+import {
+  ChatStreamCallback,
+  LLMChatResult,
+  ScoreEmbeddingsOptions,
+} from "./types";
 
 export const CHAT_COMPLETIONS_API_URL =
   "https://api.openai.com/v1/chat/completions";
